Manual Cookie Handling in Web Client
Commas and Semicolons in Values could cause trouble, and were in the case of DishNet using a JSON value with Commas.
New constants for Commas and Semicolons are included for potential future use with client-side cookie reading or manipulation.
Cookie Jars also are updated in the extended WebClient class after a send or receive, even if it receives an error.
The response handling also was getting messy, so I wrote a restructured set of functions for WebResponse events.